Reworked logging of connection information
Insteading of passing parseConn() a bunch of values it doesn't actually need to use,
it has been modified to fill out the structs that record the connection information
and then those structs are used to generate the connection log.

This has the advantage of logging everything we know about a connection up to the
point where an error occured.