Fixed race condition and other threading pitfalls
The use of a boolean write variable isn't safe with threading. The
value could be changed by the goroutine after the if statement was
already checked.  Also, when the goroutine marked the writer as false,
the range Logchan log writing code didn't have a provision to wait for
it to be true so logs would get dropped while the file was being
changed by the goroutine.

Instead I've used a mutex which provides the necessary synchronization
so that nothing can get lost and there aren't any race conditions.