Simplified worker / reader channel model
Previously each worker had their own work channel and there was a
channel that took worker channels and a goroutine that took work and
put it onto one of the worker's channels.  This commit eliminates this
extra "channel of channels" and lets all the workers share one work
channel.  Benchmarks doesn't show this to be any faster or slower than
the previous version but considering it's simpler, it seems like the
way to go.

In previous versions of Go it may have made more sense to reduce the
number of goroutines blocked reading a channel but with all of the
channel performance improvements that have been made there isn't any
compelling reason to go with the channel of channels dispatcher method
anymore.